+++ app/assets/javascripts/application.js
@@ -14,6 +14,7 @@
 //= require activestorage
 //= require turbolinks
 //= require jquery3
+//= require jquery-ui/widgets/autocomplete
 //= require toastr
 //= require bootstrap
 //= require_tree .