+++ app/chewy/customers_index.rb
@@ -1,0 +1,7 @@
+class CustomersIndex < Chewy::Index
+  define_type Customer.active do
+    field :name, :code, :abbr_name, :phone
+  end
+
+
+end