+++ app/chewy/products_index.rb
@@ -1,0 +1,2 @@
+class ProductsIndex < Chewy::Index
+end