+++ app/chewy/suppliers_index.rb
@@ -1,0 +1,2 @@
+class SuppliersIndex < Chewy::Index
+end