+++ db/migrate/20180525102218_change_status_default_value.rb
@@ -1,0 +1,7 @@
+class ChangeStatusDefaultValue < ActiveRecord::Migration[5.2]
+  def change
+    change_column :products, :status, :integer, :default => 0
+    change_column :customers, :status, :integer, :default => 0
+    change_column :suppliers, :status, :integer, :default => 0
+  end
+end